タスクを削除・編集できるのはcurrent_userをif文で

+++ app/controllers/tasks_controller.rb
@@ -29,18 +29,26 @@ class TasksController < ApplicationController
   def update
     board = Board.find(params[:board_id])
     @task = board.tasks.find(params[:id])
-    if @task.update(task_params)
-      redirect_to board_task_path(@task.board), notice: '更新しました'
-    else
-      render :edit
-    end
+      if @task.user_id == current_user.id
+          if @task.update(task_params)
+            redirect_to board_task_path(@task.board), notice: '更新しました'
+          else
+            render :edit
+          end
+      else
+        redirect_to board_task_path
+      end
   end
 
   def destroy
     board = Board.find(params[:board_id])
     task = board.tasks.find(params[:id])
-    task.destroy!
-    redirect_to root_path
+    if task.user_id == current_user.id
+      task.destroy!
+      redirect_to root_path
+    else
+      redirect_to board_task_path
+    end
   end
 
   private
